// types/portfolio.ts
export type PortfolioLocalImage = {
  uri: string;
};

export type Portfolio = {
  _id?: string;
  profilePhoto?: string | PortfolioLocalImage;
  name: string;
  profession: string;
  bio: string;
  location: string;
  contact: {
    countryCode: string;
    phone: string;
  };
  email: string;

  services: string[];
  skills: string[];
  id?: string;
  userId?: string;
  rating?: number;
  experience?: any[];
  gallery: (string | PortfolioLocalImage)[];

  links: {
    linkedin?: string;
    github?: string;
    website?: string;
    [key: string]: string | undefined;
  };
};